pyproject.toml
[tool.poetry]
name = "parabois-bot"
version = "0.0.1"
description = "Parabois discord bot"
authors = ["ZerdoX-x <zerdox.cool4@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
discord = "^1.7.3"
python-dotenv = "^0.19.0"
Flask = "^2.0.2"
waitress = "^2.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
